Check instructuion cache on read instead of write
Current implementation invalidates cache of decoded instruction if
memory cell is rewritten. However, we are going to introduce methods to
overwrite memory externally (GDB and external I/O), and it becomes too
hard to track whether memory is updated. Instead, we will read memory
bytes and cache of decoded instructions and decide whether redecoding is
needed.